backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
gunicorn>=22.0
openai>=1.30
python-dotenv>=1.0
slowapi>=0.1.9
